backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-cloud-bigquery>=3.13.0
python-dotenv>=1.0.0
structlog>=23.2.0
scikit-learn>=1.3.0\nnumpy>=1.24.0\nprophet>=1.1.5
pandas>=2.0.0
